% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_read_if_filepath}
\alias{db_read_if_filepath}
\title{Parse Signature Database.}
\usage{
db_read_if_filepath(db, dbtype = "A")
}
\arguments{
\item{db}{a signature collection}

\item{dbtype}{a string describing the db type (e.g. 'db_sbs', 'db_sv', etc)}
}
\value{
sigminer formatted signature collection if db is a string, else returns db
}
\description{
If db is a string, assumes it is a path to a csv_tidy formatted file and parses it into a sigminer (wide) format.
If db is not a string, returns db.
}
