% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_tally}
\alias{fix_tally}
\title{Longify tally matrices}
\usage{
fix_tally(matrix)
}
\arguments{
\item{matrix}{tally matrix prepared by \code{\link[=prepare_matrix]{prepare_matrix()}}}
}
\value{
tibble with long tally data
}
\description{
Longify tally matrices
}
