% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{postprocess_tally_matrix}
\alias{postprocess_tally_matrix}
\title{Post-process a tally matrix.}
\usage{
postprocess_tally_matrix(mx, class, samples)
}
\arguments{
\item{mx}{A numeric matrix, or \code{NULL}. The input tally matrix to be processed.}

\item{class}{A character vector specifying the class of mutations.}

\item{samples}{An integer specifying the number of samples.}
}
\value{
A numeric matrix with dimensions \verb{channels × samples}.
}
\description{
This function performs two operations on the input matrix \code{mx}:
}
\details{
\enumerate{
\item If \code{mx} is \code{NULL} (which can occur when the supplied MAF file doesn't contain any mutations
of the relevant class), the function produces a default numeric matrix of size \verb{samples × channels}
where all values are 0.

\item The matrix is transposed to become a \verb{channels × samples} matrix.
}
}
