% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureDatabaseCreation.R
\name{sig_add_to_database}
\alias{sig_add_to_database}
\title{Add data to sqlite DB}
\usage{
sig_add_to_database(
  signature_directory,
  sqlite_db,
  ref = c("hg19", "hg38"),
  metadata = NULL
)
}
\arguments{
\item{signature_directory}{path to the directory produced by \code{\link[=sig_analyse_mutations]{sig_analyse_mutations()}}}

\item{sqlite_db}{path to the sqlite database produced by \code{\link[=sig_create_database]{sig_create_database()}}}

\item{ref}{reference genome: one of hg19 or hg38 (string)}

\item{metadata}{a sample-level data.frame with columns sampleId, disease and description. If not supplied will be automatically generated.}
}
\value{
invisible(NULL). This function is run for its side effects
}
\description{
Load signature data into sqlite DB
}
