% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalysis.R
\name{sig_analyse_mutations_single_sample_from_files}
\alias{sig_analyse_mutations_single_sample_from_files}
\title{Mutational Signature Analysis}
\usage{
sig_analyse_mutations_single_sample_from_files(
  sample_id,
  vcf_snv = NULL,
  segment = NULL,
  vcf_sv = NULL,
  pass_only = TRUE,
  structuralvariant = NULL,
  exclude_sex_chromosomes = TRUE,
  allow_multisample = TRUE,
  db_sbs = NULL,
  db_indel = NULL,
  db_dbs = NULL,
  db_cn = NULL,
  db_sv = NULL,
  ref_tallies = NULL,
  ref = c("hg38", "hg19"),
  output_dir = "./signatures",
  exposure_type = c("absolute", "relative"),
  n_bootstraps = 100,
  temp_dir = tempdir(),
  cores = future::availableCores()
)
}
\arguments{
\item{sample_id}{string representing the tumour sample identifier (in your VCFs and other files).}

\item{vcf_snv}{A character string specifying the path to the VCF file.}

\item{pass_only}{A logical value indicating whether to filter variants to include only those with a "PASS" filter status. Default is TRUE.}

\item{structuralvariant}{The input structural variant data.frame. See \code{\link[sigminer:read_sv_as_rs]{sigminer::read_sv_as_rs()}} for details}

\item{allow_multisample}{Do NOT throw an error for VCFs with >2 samples.
Added since some tumor-normal pipelines add a third sample describing RNA genotypes of mutations found in DNA and these 3-sample VCFs would fail purely because they have 3 samples present.
Whether TRUE/FALSE, this function still assumes every mutation in the VCF relates to the \code{sample_id} supplied, so please exclude any homozygous ref alleles for your sample of interest upstream.}

\item{db_sbs, db_indel, db_dbs, db_cn}{a signature collection data.frame where rows are channels and columns are signatures.}

\item{ref_tallies}{path to a parquet file describing catalogues of a reference database. Can be produced from a folder full of sigminerUtils signature outputs using \code{\link[=sig_create_reference_set]{sig_create_reference_set()}}.
If building yourself, it must contain columns class,sample,channel,type,fraction,count. If building your own, we recommend partitioning on class then sample.}

\item{ref}{A character vector specifying the reference genome. One of 'hg38' or 'hg19'.}

\item{output_dir}{The output directory for storing results. Default is "./signatures".}

\item{exposure_type}{The type of exposure. Can be "absolute" or "relative". One of "absolute" or "relative"}

\item{n_bootstraps}{The number of bootstrap iterations for fitting signatures. Default is 100.}

\item{temp_dir}{The temporary directory for storing intermediate files. Default is tempdir().}

\item{cores}{Number of cores to use.}
}
\value{
None.
}
\description{
Run all signature mutation analyses possible from file inputs.
}
