% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureDatabaseCreation.R
\name{sig_create_database}
\alias{sig_create_database}
\title{Create a Signature Database}
\usage{
sig_create_database(sqlite_db, overwrite = TRUE)
}
\arguments{
\item{sqlite_db}{path to sql db}

\item{overwrite}{overwrite old sqlite db}
}
\value{
invisible(NULL) run for its side effects
}
\description{
Create a Signature Database
}
