% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateReferenceSet.R
\name{sig_create_reference_set}
\alias{sig_create_reference_set}
\title{Signature analysis outputs into a reference set}
\usage{
sig_create_reference_set(path_to_signature_directory = "signatures/")
}
\description{
Point to a folder full of outputs from \strong{sigminerUtils::sig_analyse_mutations()} and create
}
