% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_cosine_similarity.R
\name{sig_pairwise_similarity}
\alias{sig_pairwise_similarity}
\title{Compute Pairwise Similarity Table from Database}
\usage{
sig_pairwise_similarity(
  path_db,
  cores = future::availableCores(),
  overwrite = FALSE,
  assume_sensible_input = TRUE,
  verbose = TRUE,
  low_ram = FALSE
)
}
\arguments{
\item{path_db}{path to sqlite database}

\item{cores}{number of cores to use}

\item{overwrite}{should we only calculate similarity for pairs not already in the pairwiseSimilarity table?}

\item{assume_sensible_input}{should we speed up cosine sim by ~30x by assuming input we supply will be appropriate (channels in the same order)}

\item{verbose}{verbose}

\item{low_ram}{run in a low-ram version where instead of loading all decompositions for all samples into memory, we repeatedly query the sqlite db to fetch each sample}
}
\value{
a data.frame with pairwise cosine similarities (sample 1 & sample 2 columns based on R 'sort' order of sampleIds & upper triangle only)
}
\description{
Compute Pairwise Similarity Table from Database
}
