% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigminerToSigstory.R
\name{sigminer2sigstory}
\alias{sigminer2sigstory}
\title{Create Sigstory Inputs from Sigminer Outputs}
\usage{
sigminer2sigstory(
  signature_folder = "colo829_signature_results/COLO829v003T",
  p_threshold = 0.5
)
}
\value{
A list with all the information required to build a signature report
}
\description{
This function takes the outputs of sigminerUtils and parses them into sigstory ready visualisations & analysies
}
