% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing_models.R
\name{sigminer_tally_to_sigverse_catalogue_collection}
\alias{sigminer_tally_to_sigverse_catalogue_collection}
\title{Sigminer tally matrix -> a sigverse collection}
\usage{
sigminer_tally_to_sigverse_catalogue_collection(matrix, class)
}
\arguments{
\item{matrix}{sigminer tally matrix (must be transposed)}
}
\value{
sigverse catalogue collection. See \code{\link[sigshared:catalogue_collections]{sigshared::example_catalogue_collection()}}
}
\description{
Sigminer tally matrix -> a sigverse collection
}
