% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigminerUtils-package.R
\docType{package}
\name{sigminerUtils-package}
\alias{sigminerUtils}
\alias{sigminerUtils-package}
\title{sigminerUtils: Utilities for Running Signature Analysis using Sigminer}
\description{
A collection of utilities for running signature analysis with sigminer. Results are stored in a sqlite database and indexed for downstream analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selkamand/sigminerUtils}
  \item Report bugs at \url{https://github.com/selkamand/sigminerUtils/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

}
\keyword{internal}
