% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tally}
\alias{tally}
\title{Tally}
\usage{
tally(object, mode, ref_genome, cores)
}
\arguments{
\item{object}{a \link[sigminer]{CopyNumber} object or \link[sigminer]{MAF} object or SV object (from \link[sigminer]{read_sv_as_rs}).}

\item{mode}{type of mutation matrix to extract, can be one of 'SBS', 'DBS' and 'ID'.}

\item{ref_genome}{'BSgenome.Hsapiens.UCSC.hg19', 'BSgenome.Hsapiens.UCSC.hg38',
'BSgenome.Mmusculus.UCSC.mm10',  'BSgenome.Mmusculus.UCSC.mm9', etc.}

\item{cores}{number of computer cores to run this task.
You can use \code{\link[future:re-exports]{future::availableCores()}} function to check how
many cores you can use.}
}
\value{
a \code{list} contains a \code{matrix} used for NMF de-composition.
}
\description{
A wrapper for \code{\link[sigminer:sig_tally]{sigminer::sig_tally()}} that adds 2 features.
\enumerate{
\item If supplied with a MAF with none of the type of mutation for the given 'MODE' instead of erroring returns an empty matrix.
}
}
