% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_compressed_csv}
\alias{write_compressed_csv}
\title{Write compressed CSV}
\usage{
write_compressed_csv(x, file)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame.
    If not, it is attempted to coerce \code{x} to a data frame.}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing.  \code{""} indicates output to the console.}
}
\description{
Write compressed CSV
}
