% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateReferenceSet.R
\name{parse_sig_files}
\alias{parse_sig_files}
\title{Parse Signature Analysis Output Files}
\usage{
parse_sig_files(x, colClasses = NA)
}
\arguments{
\item{x}{paths to any file produced by sig_analyse_mutations (character vector)}
}
\value{
a data.frame with 1 row per path, that includes filename-encoded metadata and a list-column with dataframes (contents)
}
\description{
Parse Signature Analysis Output Files
}
