% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureDatabaseCreation.R
\name{sig_add_to_database}
\alias{sig_add_to_database}
\title{Add data to sqlite DB}
\usage{
sig_add_to_database(signature_directory, sqlite_db)
}
\arguments{
\item{signature_directory}{path to the directory produced by \code{\link[=sig_analyse_mutations]{sig_analyse_mutations()}}}

\item{sqlite_db}{path to the sqlite database produced by \link{(sig_create_database)}}
}
\value{
invisible(NULL). This function is run for its side effects
}
\description{
Load signature data into sqlite DB
}
