% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalysis.R
\name{sig_analyse_mutations}
\alias{sig_analyse_mutations}
\title{Mutational Signature Analysis}
\usage{
sig_analyse_mutations(
  maf,
  copynumber = NULL,
  structuralvariant = NULL,
  db_sbs = NULL,
  db_indel = NULL,
  db_dbs = NULL,
  db_cn = NULL,
  db_sv = NULL,
  ref_tallies = NULL,
  ref = c("hg38", "hg19"),
  output_dir = "./signatures",
  exposure_type = c("absolute", "relative"),
  n_bootstraps = 100,
  temp_dir = tempdir(),
  cores = future::availableCores()
)
}
\arguments{
\item{maf}{The input MAF file. Can be a maf object or the path to a MAF file}

\item{copynumber}{The input copynumber data.frame. See \code{\link[sigminer:read_copynumber]{sigminer::read_copynumber()}} for details.}

\item{structuralvariant}{The input structural variant data.frame. See \code{\link[sigminer:read_sv_as_rs]{sigminer::read_sv_as_rs()}} for details}

\item{db_sbs, db_indel, db_dbs, db_cn, db_sv}{a signature collection data.frame where rows are channels and columns are signatures. Row names should be signature channels. See \code{sigstash::sig_load("COSMIC_v3.4_SBS_GRCh38", format = "sigminer")}  for an example. Alternatively, you can supply a path to signature collections in tidy_csv format (see \code{\link[sigstash:sig_read_signatures]{sigstash::sig_read_signatures()}} for details)}

\item{ref_tallies}{path to a parquet file describing catalogues of a reference database. Can be produced from a folder full of sigminerUtils signature outputs using \code{\link[=sig_create_reference_set]{sig_create_reference_set()}}.
If building yourself, it must contain columns class,sample,channel,type,fraction,count. If building your own, we recommend partitioning on class then sample.}

\item{ref}{A character vector specifying the reference genome. One of 'hg38' or 'hg19'.}

\item{output_dir}{The output directory for storing results. Default is "./signatures".}

\item{exposure_type}{The type of exposure. Can be "absolute" or "relative". One of "absolute" or "relative"}

\item{n_bootstraps}{The number of bootstrap iterations for fitting signatures. Default is 100.}

\item{temp_dir}{The temporary directory for storing intermediate files. Default is tempdir().}

\item{cores}{Number of cores to use.}
}
\value{
None.
}
\description{
Run all signature mutation analyses possible from MAF inputs on
}
