% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureAnalysis.R
\name{sig_analyse_mutations}
\alias{sig_analyse_mutations}
\title{Mutational Signature Analysis}
\usage{
sig_analyse_mutations(
  maf,
  somatic_ids,
  db_sbs = NULL,
  db_indel = NULL,
  db_dbs = NULL,
  ref = c("hg38", "hg19"),
  output_dir = "./signatures",
  exposure_type = c("absolute", "relative"),
  n_bootstraps = 100,
  temp_dir = tempdir()
)
}
\arguments{
\item{maf}{The input MAF file.}

\item{somatic_ids}{A vector of sample IDs for somatic mutations.}

\item{ref}{A character vector specifying the reference genome. One of 'hg38' or 'hg19'.}

\item{output_dir}{The output directory for storing results. Default is "./signatures".}

\item{exposure_type}{The type of exposure. Can be "absolute" or "relative". One of "absolute" or "relative"}

\item{n_bootstraps}{The number of bootstrap iterations for fitting signatures. Default is 100.}

\item{temp_dir}{The temporary directory for storing intermediate files. Default is tempdir().}
}
\value{
None.
}
\description{
Run all signature mutation analyses possible from MAF inputs on
}
