% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateReferenceSet.R
\name{sig_create_reference_set}
\alias{sig_create_reference_set}
\title{Signature analysis outputs into a reference set}
\usage{
sig_create_reference_set(
  path_to_signature_directory = "signatures/",
  outfolder = "reference",
  format = c("parquet", "csv.gz"),
  create_umaps = TRUE,
  umap_n_neighbours = 15,
  seed = 111,
  verbose = TRUE
)
}
\arguments{
\item{path_to_signature_directory}{path to signature directory}

\item{outfolder}{path to save reference set}

\item{format}{format to save reference set as (parquet or bgzipped csv)}

\item{create_umaps}{should umap references be created (flag)}

\item{umap_n_neighbours}{This parameter sets the balance between local and global structure in UMAP.
The value determines the size of the local neighborhood UMAP which is used to learn the manifold structure of the data.
Lower values focus on local details. higher values emphasise more global patterns. Must be less than the number of samples in your dataset.}

\item{verbose}{verbose (flag)}
}
\value{
this function is run for its side effects. Invisibly returns NULL.
}
\description{
Point to a folder full of outputs from \strong{sigminerUtils::sig_analyse_mutations()} and create
}
