% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigminerToSigstory.R
\name{sigminer2sigstory}
\alias{sigminer2sigstory}
\title{Create Sigstory Inputs from Sigminer Outputs}
\usage{
sigminer2sigstory(
  signature_folder = "colo829_signature_results_with_refset/COLO829v003T",
  rds_outfile = "sigstory.rds",
  ref_tallies = NULL,
  ref_exposures = NULL,
  ref_metadata = NULL,
  sparsity_pvalue = 0.05
)
}
\arguments{
\item{signature_folder}{folder containing signature data}

\item{rds_outfile}{file to store serialised signature results that can serve as sigstory input.}

\item{sparsity_pvalue}{max p-value, below which we consider the signatures to be stable across bootstraps. See \code{\link[sigstats:sig_compute_experimental_p_value]{sigstats::sig_compute_experimental_p_value()}} for pvalue computation.}
}
\value{
A list with all the information required to build a signature report
}
\description{
This function takes the outputs of sigminerUtils and parses them into sigstory ready visualisations & analysies
}
